// vendor/raw.githubusercontent.com/#mikaelporttila_0c368/rss/refs/tags/1.0.3/src/types/fields/dublin_core_fields.ts
export enum DublinCoreFields {
  Contributor = "dc:contributor",
  Coverage = "dc:coverage",
  Creator = "dc:creator",
  Date = "dc:date",
  DateRaw = "dc:dateRaw",
  Description = "dc:description",
  Format = "dc:format",
  Language = "dc:language",
  Identifier = "dc:identifier",
  Publisher = "dc:publisher",
  Relation = "dc:relation",
  Rights = "dc:rights",
  Source = "dc:source",
  Subject = "dc:subject",
  Title = "dc:title",
  Type = "dc:type",
  Abstract = "dc:abstract",
  AccessRights = "dc:accessRights",
  AccrualMethod = "dc:accrualMethod",
  AccrualPeriodicity = "dc:accrualPeriodicity",
  AccrualPolicy = "dc:accrualPolicy",
  Alternative = "dc:alternative",
  Audience = "dc:audience",
  Available = "dc:available",
  BibliographicCitation = "dc:bibliographicCitation",
  ConformsTo = "dc:conformsTo",
  Created = "dc:created",
  CreatedRaw = "dc:createdRaw",
  DateAccepted = "dc:dateAccepted",
  DateAcceptedRaw = "dc:dateAcceptedRaw",
  Copyrighted = "dc:copyrighted",
  CopyrightedRaw = "dc:CopyrightedRaw",
  DateSubmitted = "dc:dateSubmitted",
  DateSubmittedRaw = "dc:dateSubmittedRaw",
  EducationLevel = "dc:educationLevel",
  Extent = "dc:extent",
  HasFormat = "dc:hasFormat",
  HasPart = "dc:hasPart",
  HasVersion = "dc:hasVersion",
  InstructionalMethod = "dc:instructionalMethod",
  IsFormatOf = "dc:isFormatOf",
  IsPartOf = "dc:isPartOf",
  IsReferencedBy = "dc:isReferencedBy",
  IsReplacedBy = "dc:isReplacedBy",
  IsRequiredBy = "dc:isRequiredBy",
  Issued = "dc:issued",
  IsVersionOf = "dc:isVersionOf",
  License = "dc:license",
  Mediator = "dc:mediator",
  Medium = "dc:medium",
  Modified = "dc:modified",
  ModifiedRaw = "dc:modifiedRaw",
  Provenance = "dc:provenance",
  References = "dc:references",
  Replaces = "dc:replaces",
  Requires = "dc:requires",
  RightsHolder = "dc:rightsHolder",
  Spatial = "dc:spatial",
  TableOfContents = "dc:tableOfContents",
  Tempora = "dc:tempora",
  Valid = "dc:valid",
  Box = "dc:Box",
  DCMIType = "dc:DCMIType",
  DDC = "dc:DDC",
  IMT = "dc:IMT",
  ISO3166 = "dc:ISO3166",
  ISO6392 = "dc:ISO639-2",
  LLC = "dc:LLC",
  LCSH = "dc:LCSH",
  MESH = "dc:MESH",
  NLM = "dc:NLM",
  Period = "dc:Period",
  Point = "dc:Point",
  RFC1766 = "dc:RFC1766",
  RFC3066 = "dc:RFC3066",
  TGN = "dc:TGN",
  UDC = "dc:UDC",
  URI = "dc:URI",
  W3CDTF = "dc:W3CDTF",
}